+++ src/app/store.js
@@ -1,6 +1,6 @@
 // used to pass data more easily between react components
 import { configureStore } from '@reduxjs/toolkit'
-import counterReducer from '../features/counter/counterSlice'
+import counterReducer from '../components/counter/counterSlice'
 
 export default configureStore({
   reducer: {
