+++ src/pages/homepage.js
@@ -4,7 +4,7 @@ import React, {useState, useId} from "react";
 import Popup from 'reactjs-popup';
 import PlayerWaiting from "./PlayerWaiting.js";
 import { useSelector, useDispatch } from 'react-redux'
-import { setValue } from '../features/counter/counterSlice.js'
+import { setValue } from '../components/counter/counterSlice.js'
 import background from "../images/stars-background.jpg";
 
 import { useNavigate } from 'react-router-dom';
